Создал компонент компонент Список предложений неподалёку
